flask_api_project/requirements.txt
Flask
mysql-connector-python
werkzeug
pyjwt
dotenv